task_creator/requirements.txt
Flask==1.0.2
Flask-RESTful==0.3.7
pandas==0.25.1
XlsxWriter==1.1.9
sendgrid==6.0.5
google-cloud-tasks==1.2.1
